+++ src/app/assortment/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
 
-import { Assortment as AssortmentSection } from '@/containers/Assortment/Assortment'
+import { Assortment as AssortmentSection } from '@/containers/Assortment'
 import { strings } from '@/utils/strings'
 
 export const metadata: Metadata = {
