+++ src/app/ordering/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
 
-import { Ordering as OrderingSection } from '@/containers/Ordering/Ordering'
+import { Ordering as OrderingSection } from '@/containers/Ordering'
 import { strings } from '@/utils/strings'
 
 export const metadata: Metadata = {
