+++ src/components/Gallery/Image/index.ts
@@ -1,0 +1,1 @@
+export { GalleryImage, type GalleryImageProps } from './Image'
