+++ src/components/Gallery/index.ts
@@ -1,0 +1,1 @@
+export { Gallery } from './Gallery'
