+++ src/components/Lightbox/index.ts
@@ -1,0 +1,1 @@
+export { Lightbox } from './Lightbox'
