+++ src/components/ServiceWrapper/ServiceItem/index.ts
@@ -1,0 +1,1 @@
+export { type ServiceItemProps, ServiceItem } from './ServiceItem'
