+++ src/components/ServiceWrapper/index.ts
@@ -1,0 +1,1 @@
+export { Service } from './Service'
