+++ src/components/SocialMediaLinkWrapper/index.ts
@@ -1,0 +1,1 @@
+export { SocialMediaLinkWrapper } from './SocialMediaLinkWrapper'
