+++ src/components/TestimonialWrapper/TestimonialItem/index.ts
@@ -1,0 +1,1 @@
+export { type TestimonialItemProps, TestimonialItem } from './TestimonialItem'
