+++ src/components/TestimonialWrapper/index.ts
@@ -1,0 +1,1 @@
+export { Testimonial } from './Testimonial'
