+++ src/containers/Assortment/index.ts
@@ -1,0 +1,1 @@
+export { Assortment } from './Assortment'
