+++ src/containers/Contact/index.ts
@@ -1,0 +1,1 @@
+export { Contact } from './Contact'
