+++ src/containers/Ordering/index.ts
@@ -1,0 +1,1 @@
+export { Ordering } from './Ordering'
